pubmed_tool/pyproject.toml
[project]
name = "pubmed-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "Supriya",email = "supriyawakchaure0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "typer[all] (>=0.16.0,<0.17.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "streamlit (>=1.47.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "pubmed_tool", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
